Refactored userSelection state by creating a specify object.
Created a nested specify object to hold the various types of specifications the user could select in a more organized way. In order to do this I had to refactor many items in order to ensure everything was still properly working with the new structure. I did this to not only keep it more organized, but hopefully ensure an easier way to do checkbox form validation.